+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog + Dev Log
 
+### (Feb 8, 2025)
+
+- Fixed files not displaying when the main selected folder is a hidden folder (aka starts with '.')
+- Route history resets when changing notes folder now
+- Fixed navigation history when you enter into trash or archive while not on main home screen
+
 ### (Feb 7, 2025)
 
 **Changes:**